+++ backend/package.json
@@ -5,6 +5,7 @@
   "license": "MIT",
   "dependencies": {
     "@google-cloud/text-to-speech": "^3.4.0",
+    "amd-loader": "^0.0.8",
     "express": "^4.17.2"
   },
   "scripts": {
