+++ backend/src/amd.js
@@ -1,0 +1,3 @@
+define(function (require, exports, module) {
+  exports.B = "B";
+});
