+++ backend/src/controllers/translate.ts
@@ -1,0 +1,3 @@
+import { speech } from "../utils";
+
+export const getTranslate = (req: Express.Request, res: Express.Response) => {};
