+++ backend/src/routes/translate.ts
@@ -1,0 +1,6 @@
+import express from "express";
+import { getTranslate } from "../controllers";
+
+export const translateRouter = express.Router();
+
+translateRouter.get("/", getTranslate);
