+++ backend/src/server.ts
@@ -1,16 +1,14 @@
-const express = require("express");
-const bodyParser = require("body-parser")
-
+import express from "express";
+import bodyParser from "body-parser";
+import { translateRouter as translate } from "./routes";
 
 const PORT = 4000;
-
 const app = express();
+
 app.use(bodyParser.json());
 
-app.get("/", () => {
-  console.log("request has been made to '/'")
-})
+app.use("/translate", translate);
 
-app.listen(PORT,() => {
-  console.log(`app is running on PORT ${PORT}`)
-} )
+app.listen(PORT, () => {
+  console.log(`app is running on PORT ${PORT}`);
+});
